Models - Change the primary keys to use the ebird codes and identifiers
This was not done before as having varchars as a primary key seemed like
an enormous cost - even though it isn't, as the indexes don't contain the
keys themselves. There is likely a cost in overall database size though.

The changeover was made for the ebirders-pt project where Checklist, and
Observations can be filtered on country, state, observer, etc. The URLs
contain the codes, and to avoid joins, these are converted to the integer
prinmary keys to filter the tables. Changing to the codes and identifiers
eliminates this step, so the code is simpler and cleaner.