APILoader - Refactor fetching the Observer for a checklist
The eBird API only contains the name of the observer - which may not be
unique. To avoid having all the observations made by Jane Doe assigned
to one person, when an observer is added, the web page for the checklist
is scraped to get the eBird identifier, e.g. USER12345678. We don't want
to scrape the web page for every checklist, so the assumption is that
every name is unique, until it is discovered it is not.

Generally this is not a problem as people pick names so they are
unique.  Nobody wants somebody else to get the credit for birds they
saw. However "Anonymous eBirder" presents a problem. This is the
generic name given to any observer who wanted to remain anonymous. If
you are just browsing the data then this is not a problem. However if
you start processing the data to report on number of species seenm or
number of checklists submitted, then "Anonymous eBirder" rises to the
top of the lists.

To "solve" this the 'multiple' attribute was added to Observer. When
this flag is set, then the checklist web page is always scraped, so
the correct user can be found. The code to do this worked, but not
well enough and, mostly for historical reaasons, as the code evolved,
some Observers with duplicate identifiers got added to the database.
That wasn't a problem until the primary key change was made. This
commit tries to clean up the get_observer() method to avoid creating
an Observer where there is aleady one with the same identifier, since
that will obviously now fail.

The problem is not solved, however. If someone changes their name, then
the APILoader will try to add a new Observer with an existing identifier.
This will only be solved when eBird adds the user identifier to the
data returned by the API.