Models - Add timestamp for when a record was added to the database
Timestamp fields, created and modified are usually added by default to
models, and usually they are not needed, so this time they were left
out. Now there is finally a need to start analysing the data, so the
created timestamp was added. For tracking changes to the data an event
table will probably be used.